realtime/python/requirements.txt
numpy
websockets 
loguru
python-dotenv